+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [HexaPDF::Revision#reset_objects] for deleting all live loaded and added
   objects
+- Support for removing or flattening annotations to the `hexapdf modify` command
 
 ### Changed